+++ RTWWebServer/DTOs/UserSession.cs
@@ -1,7 +1,7 @@
 namespace RTWWebServer.DTOs;
 
-public class UserSession(int userId, string token)
+public class UserSession(long userId, string token)
 {
-    public int UserId { get; set; } = userId;
+    public long UserId { get; set; } = userId;
     public string Token { get; set; } = token;
 }
